requirements.txt
slack-bolt>=1.18.0
python-dotenv>=0.19.0 